Revert households ≠ claim codes (every household has a code)
Reverts PR #254 (merge 3f5f03c4) per product-owner decision (2026-07-15):
there is no code-less household — families.public_id is required + globally
unique again.

Migration approach: KEEP 0032 (it already ran on prod D1 + is tracked in
d1_migrations; deleting it would desync a fresh D1) and ADD a forward
reversing migration 0033_households_require_code.sql that rebuilds families
back to public_id text NOT NULL + full column-level UNIQUE (dropping 0032's
partial index) via the same __keep_* FK-preserving idiom. Row ids copied
verbatim so the cascade subtree (guests/sessions/guest_events/rsvps/
guest_account_links) keeps every FK — zero orphans, proven by
migration-0033.test.ts. 0033 fails loud on a NULL public_id (no coercion).

Reverted code: deleted services/households.ts + services/issue-invite.ts
(+ tests), schemas/household.ts, the household/issue-invite routes + their
tests, the household/issue metrics, and the app.ts mount; family-deactivate
no longer special-cases code-less households; claim schema + guests-store
type back to non-null; GuestTable back to grouping by publicId (no issue-invite
UI). schema.ts + setup.ts DDL restored to NOT NULL + full unique — T-S1 green
for 0001…0033. Import auto-mint unchanged. Deleted migration-0032.test.ts
(asserted the transient nullable state 0033 undoes).

Docs: platform-plan §3.2 marked REVERSED; todo/platform PR 4 un-ticked;
todo/db 0033 entry added.
