+++ cire/organiser/src/lib/guests-store.ts
@@ -20,9 +20,7 @@ import { type Accessor, createSignal, type Setter } from "solid-js";
  *  the table dedupes to households). */
 export interface OrganiserGuestRow {
   familyId: string;
-  /** The household's claim code, or `null` for a CODE-LESS household (PR 4 —
-   *  no claimable invite until one is issued). */
-  publicId: string | null;
+  publicId: string;
   familyName: string;
   firstName: string;
   lastName: string;
